Add input_layer param to encoder
